changed quantities to Int and renamed quantities referring to kg and ml to size

+++ composeApp/src/main/kotlin/com/example/a62550_foodapp/model/ShoppingListItem.kt
@@ -4,7 +4,7 @@ package com.example.a62550_foodapp.model
 data class ShoppingListItem(
     val shopping_list_id: Int,
     val item_id: Int,
-    val quantity: Float,
+    val quantity: Int,
     val is_checked: Boolean,
     val label: String
 )
